Fixed general warnings in all classes
git-svn-id: https://svn.jboss.org/repos/jbossas/projects/specs/trunk/jboss-rmi-api_1.0_spec@111478 84be2c1e-ba19-0410-b317-a758671a6fc1

+++ src/main/java/org/jboss/com/sun/corba/se/impl/encoding/CodeSetComponentInfo.java
@@ -40,42 +40,47 @@ import java.util.StringTokenizer;
 import org.jboss.com.sun.corba.se.impl.logging.ORBUtilSystemException;
 import org.jboss.com.sun.corba.se.spi.logging.CORBALogDomains;
 
-public final class CodeSetComponentInfo {
+public final class CodeSetComponentInfo
+{
 
     /**
-     * CodeSetComponent is part of an IOR multi-component profile.  Two
-     * instances constitute a CodeSetComponentInfo (one for char and one
-     * for wchar data)
+     * CodeSetComponent is part of an IOR multi-component profile. Two instances constitute a CodeSetComponentInfo (one
+     * for char and one for wchar data)
      */
-    public static final class CodeSetComponent {
+    public static final class CodeSetComponent
+    {
         int nativeCodeSet;
+
         int[] conversionCodeSets;
 
-        public boolean equals( Object obj )
+        public boolean equals(Object obj)
         {
             if (this == obj)
-                return true ;
+                return true;
 
             if (!(obj instanceof CodeSetComponent))
-                return false ;
+                return false;
 
-            CodeSetComponent other = (CodeSetComponent)obj ;
+            CodeSetComponent other = (CodeSetComponent) obj;
 
-            return (nativeCodeSet == other.nativeCodeSet) &&
-                Arrays.equals( conversionCodeSets, other.conversionCodeSets ) ;
+            return (nativeCodeSet == other.nativeCodeSet)
+                    && Arrays.equals(conversionCodeSets, other.conversionCodeSets);
         }
 
         public int hashCode()
         {
-            int result = nativeCodeSet ;
-            for (int ctr=0; ctr<conversionCodeSets.length; ctr++)
-                result = 37*result + conversionCodeSets[ctr] ;
-            return result ;
+            int result = nativeCodeSet;
+            for (int ctr = 0; ctr < conversionCodeSets.length; ctr++)
+                result = 37 * result + conversionCodeSets[ctr];
+            return result;
         }
 
-        public CodeSetComponent() {}
+        public CodeSetComponent()
+        {
+        }
 
-        public CodeSetComponent(int nativeCodeSet, int[] conversionCodeSets) {
+        public CodeSetComponent(int nativeCodeSet, int[] conversionCodeSets)
+        {
             this.nativeCodeSet = nativeCodeSet;
             if (conversionCodeSets == null)
                 this.conversionCodeSets = new int[0];
@@ -83,7 +88,8 @@ public final class CodeSetComponentInfo {
                 this.conversionCodeSets = conversionCodeSets;
         }
 
-        public void read(MarshalInputStream in) {
+        public void read(MarshalInputStream in)
+        {
             nativeCodeSet = in.read_ulong();
             int len = in.read_long();
             conversionCodeSets = new int[len];
@@ -91,13 +97,15 @@ public final class CodeSetComponentInfo {
 
         }
 
-        public void write(MarshalOutputStream out) {
+        public void write(MarshalOutputStream out)
+        {
             out.write_ulong(nativeCodeSet);
             out.write_long(conversionCodeSets.length);
             out.write_ulong_array(conversionCodeSets, 0, conversionCodeSets.length);
         }
 
-        public String toString() {
+        public String toString()
+        {
             StringBuffer sbuf = new StringBuffer("CodeSetComponent(");
 
             sbuf.append("native:");
@@ -105,40 +113,43 @@ public final class CodeSetComponentInfo {
             sbuf.append(" conversion:");
             if (conversionCodeSets == null)
                 sbuf.append("null");
-            else {
-                for (int i = 0; i < conversionCodeSets.length; i++) {
+            else
+            {
+                for (int i = 0; i < conversionCodeSets.length; i++)
+                {
                     sbuf.append(Integer.toHexString(conversionCodeSets[i]));
                     sbuf.append(' ');
                 }
             }
-            sbuf.append( ")" ) ;
+            sbuf.append(")");
 
             return sbuf.toString();
         }
     }
 
     private CodeSetComponent forCharData;
+
     private CodeSetComponent forWCharData;
 
-    public boolean equals( Object obj )
+    public boolean equals(Object obj)
     {
         if (this == obj)
-            return true ;
+            return true;
 
         if (!(obj instanceof CodeSetComponentInfo))
-            return false ;
+            return false;
 
-        CodeSetComponentInfo other = (CodeSetComponentInfo)obj ;
-        return forCharData.equals( other.forCharData ) &&
-            forWCharData.equals( other.forWCharData ) ;
+        CodeSetComponentInfo other = (CodeSetComponentInfo) obj;
+        return forCharData.equals(other.forCharData) && forWCharData.equals(other.forWCharData);
     }
 
     public int hashCode()
     {
-        return forCharData.hashCode() ^ forWCharData.hashCode() ;
+        return forCharData.hashCode() ^ forWCharData.hashCode();
     }
 
-    public String toString() {
+    public String toString()
+    {
         StringBuffer sbuf = new StringBuffer("CodeSetComponentInfo(");
 
         sbuf.append("char_data:");
@@ -150,70 +161,85 @@ public final class CodeSetComponentInfo {
         return sbuf.toString();
     }
 
-    public CodeSetComponentInfo() {
+    public CodeSetComponentInfo()
+    {
         forCharData = CodeSetComponentInfo.JAVASOFT_DEFAULT_CODESETS.forCharData;
         forWCharData = CodeSetComponentInfo.JAVASOFT_DEFAULT_CODESETS.forWCharData;
     }
 
-    public CodeSetComponentInfo(CodeSetComponent charData,
-                                CodeSetComponent wcharData) {
+    public CodeSetComponentInfo(CodeSetComponent charData, CodeSetComponent wcharData)
+    {
         forCharData = charData;
         forWCharData = wcharData;
     }
 
-    public void read(MarshalInputStream in) {
+    public void read(MarshalInputStream in)
+    {
         forCharData = new CodeSetComponent();
         forCharData.read(in);
         forWCharData = new CodeSetComponent();
         forWCharData.read(in);
     }
 
-    public void write(MarshalOutputStream out) {
+    public void write(MarshalOutputStream out)
+    {
         forCharData.write(out);
         forWCharData.write(out);
     }
 
-    public CodeSetComponent getCharComponent() {
+    public CodeSetComponent getCharComponent()
+    {
         return forCharData;
     }
 
-    public CodeSetComponent getWCharComponent() {
+    public CodeSetComponent getWCharComponent()
+    {
         return forWCharData;
     }
 
     /**
      * CodeSetContext goes in a GIOP service context
      */
-    public static final class CodeSetContext {
+    public static final class CodeSetContext
+    {
         private int char_data;
+
         private int wchar_data;
 
-        public CodeSetContext() {}
+        public CodeSetContext()
+        {
+        }
 
-        public CodeSetContext(int charEncoding, int wcharEncoding) {
+        public CodeSetContext(int charEncoding, int wcharEncoding)
+        {
             char_data = charEncoding;
             wchar_data = wcharEncoding;
         }
 
-        public void read(MarshalInputStream in) {
+        public void read(MarshalInputStream in)
+        {
             char_data = in.read_ulong();
             wchar_data = in.read_ulong();
         }
 
-        public void write(MarshalOutputStream out) {
+        public void write(MarshalOutputStream out)
+        {
             out.write_ulong(char_data);
             out.write_ulong(wchar_data);
         }
 
-        public int getCharCodeSet() {
+        public int getCharCodeSet()
+        {
             return char_data;
         }
 
-        public int getWCharCodeSet() {
+        public int getWCharCodeSet()
+        {
             return wchar_data;
         }
 
-        public String toString() {
+        public String toString()
+        {
             StringBuffer sbuf = new StringBuffer();
             sbuf.append("CodeSetContext char set: ");
             sbuf.append(Integer.toHexString(char_data));
@@ -225,90 +251,77 @@ public final class CodeSetComponentInfo {
 
     /**
      * Our default code set scheme is as follows:
-     *
+     * 
      * char data:
-     *
-     * Native code set:  ISO 8859-1 (8-bit)
-     * Conversion sets:  UTF-8, ISO 646 (7-bit)
-     *
+     * 
+     * Native code set: ISO 8859-1 (8-bit) Conversion sets: UTF-8, ISO 646 (7-bit)
+     * 
      * wchar data:
-     *
-     * Native code set:  UTF-16
-     * Conversion sets:  UCS-2
-     *
-     * Pre-Merlin/J2EE 1.3 JavaSoft ORBs listed ISO646 for char and
-     * UCS-2 for wchar, and provided no conversion sets.  They also
-     * didn't do correct negotiation or provide the fallback sets.
-     * UCS-2 is still in the conversion list for backwards compatibility.
-     *
+     * 
+     * Native code set: UTF-16 Conversion sets: UCS-2
+     * 
+     * Pre-Merlin/J2EE 1.3 JavaSoft ORBs listed ISO646 for char and UCS-2 for wchar, and provided no conversion sets.
+     * They also didn't do correct negotiation or provide the fallback sets. UCS-2 is still in the conversion list for
+     * backwards compatibility.
+     * 
      * The fallbacks are UTF-8 for char and UTF-16 for wchar.
-     *
-     * In GIOP 1.1, interoperability with wchar is limited to 2 byte fixed
-     * width encodings since its wchars aren't preceded by a length.
-     * Thus, I've chosen not to include UTF-8 in the conversion set
-     * for wchar data.
-     *
+     * 
+     * In GIOP 1.1, interoperability with wchar is limited to 2 byte fixed width encodings since its wchars aren't
+     * preceded by a length. Thus, I've chosen not to include UTF-8 in the conversion set for wchar data.
+     * 
      */
     public static final CodeSetComponentInfo JAVASOFT_DEFAULT_CODESETS;
-    static {
-        CodeSetComponent charData
-            = new CodeSetComponent(OSFCodeSetRegistry.ISO_8859_1.getNumber(),
-                                   new int[] {
-                                       OSFCodeSetRegistry.UTF_8.getNumber(),
-                                       OSFCodeSetRegistry.ISO_646.getNumber()
-                                   });
+    static
+    {
+        CodeSetComponent charData = new CodeSetComponent(OSFCodeSetRegistry.ISO_8859_1.getNumber(), new int[]{
+                OSFCodeSetRegistry.UTF_8.getNumber(), OSFCodeSetRegistry.ISO_646.getNumber()});
 
-        CodeSetComponent wcharData
-            = new CodeSetComponent(OSFCodeSetRegistry.UTF_16.getNumber(),
-                                   new int[]
-                                   {
-                                       OSFCodeSetRegistry.UCS_2.getNumber()
-                                   });
+        CodeSetComponent wcharData = new CodeSetComponent(OSFCodeSetRegistry.UTF_16.getNumber(),
+                new int[]{OSFCodeSetRegistry.UCS_2.getNumber()});
 
         JAVASOFT_DEFAULT_CODESETS = new CodeSetComponentInfo(charData, wcharData);
     }
 
     /**
-     * Creates a CodeSetComponent from a String which contains a comma
-     * delimited list of OSF Code Set Registry numbers.  An INITIALIZE
-     * exception is thrown if any of the numbers are not known by our
-     * registry.  Used by corba.ORB init.
-     *
-     * The first number in the list is taken as the native code set,
-     * and the rest is the conversion code set list.
-     *
+     * Creates a CodeSetComponent from a String which contains a comma delimited list of OSF Code Set Registry numbers.
+     * An INITIALIZE exception is thrown if any of the numbers are not known by our registry. Used by corba.ORB init.
+     * 
+     * The first number in the list is taken as the native code set, and the rest is the conversion code set list.
+     * 
      * The numbers can either be decimal or hex.
      */
-    public static CodeSetComponent createFromString(String str) {
-        ORBUtilSystemException wrapper = ORBUtilSystemException.get(
-            CORBALogDomains.RPC_ENCODING ) ;
+    public static CodeSetComponent createFromString(String str)
+    {
+        ORBUtilSystemException wrapper = ORBUtilSystemException.get(CORBALogDomains.RPC_ENCODING);
 
         if (str == null || str.length() == 0)
-            throw wrapper.badCodeSetString() ;
+            throw wrapper.badCodeSetString();
 
         StringTokenizer stok = new StringTokenizer(str, ", ", false);
         int nativeSet = 0;
         int conversionInts[] = null;
 
-        try {
+        try
+        {
 
             // The first value is the native code set
             nativeSet = Integer.decode(stok.nextToken()).intValue();
 
             if (OSFCodeSetRegistry.lookupEntry(nativeSet) == null)
-                throw wrapper.unknownNativeCodeset( new Integer(nativeSet) ) ;
+                throw wrapper.unknownNativeCodeset(new Integer(nativeSet));
 
             List<Integer> conversionList = new ArrayList<Integer>(10);
 
             // Now process the other values as part of the
             // conversion code set list.
-            while (stok.hasMoreTokens()) {
+            while (stok.hasMoreTokens())
+            {
 
                 // decode allows us to specify hex, decimal, etc
                 Integer value = Integer.decode(stok.nextToken());
 
                 if (OSFCodeSetRegistry.lookupEntry(value.intValue()) == null)
-                    throw wrapper.unknownConversionCodeSet( value ) ;
+                    throw wrapper.unknownConversionCodeSet(value);
 
                 conversionList.add(value);
             }
@@ -316,23 +329,25 @@ public final class CodeSetComponentInfo {
             conversionInts = new int[conversionList.size()];
 
             for (int i = 0; i < conversionInts.length; i++)
-                conversionInts[i] = ((Integer)conversionList.get(i)).intValue();
+                conversionInts[i] = conversionList.get(i);
 
-        } catch (NumberFormatException nfe) {
-            throw wrapper.invalidCodeSetNumber( nfe ) ;
-        } catch (NoSuchElementException nsee) {
-            throw wrapper.invalidCodeSetString( nsee, str ) ;
+        }
+        catch (NumberFormatException nfe)
+        {
+            throw wrapper.invalidCodeSetNumber(nfe);
+        }
+        catch (NoSuchElementException nsee)
+        {
+            throw wrapper.invalidCodeSetString(nsee, str);
         }
 
-        // Otherwise return the CodeSetComponent representing
-        // the given values
+        // Otherwise return the CodeSetComponent representing the given values
         return new CodeSetComponent(nativeSet, conversionInts);
     }
 
     /**
      * Code sets for local cases without a connection.
      */
-    public static final CodeSetContext LOCAL_CODE_SETS
-        = new CodeSetContext(OSFCodeSetRegistry.ISO_8859_1.getNumber(),
-                             OSFCodeSetRegistry.UTF_16.getNumber());
+    public static final CodeSetContext LOCAL_CODE_SETS = new CodeSetContext(OSFCodeSetRegistry.ISO_8859_1.getNumber(),
+            OSFCodeSetRegistry.UTF_16.getNumber());
 }
