+++ src/main/java/org/jboss/com/sun/corba/se/impl/protocol/giopmsgheaders/LocateReplyMessage_1_0.java
@@ -36,26 +36,29 @@ import org.omg.CORBA.SystemException;
 
 /**
  * This implements the GIOP 1.0 LocateReply header.
- *
+ * 
  * @author Ram Jeyaraman 05/14/2000
  */
 
-public final class LocateReplyMessage_1_0 extends Message_1_0
-        implements LocateReplyMessage {
+public final class LocateReplyMessage_1_0 extends Message_1_0 implements LocateReplyMessage
+{
 
     // Instance variables
 
-    private int request_id = (int) 0;
-    private int locate_status = (int) 0;
+    private int request_id = 0;
+
+    private int locate_status = 0;
+
     private IOR ior = null;
 
     // Constructors
 
-    LocateReplyMessage_1_0(ORB orb) {
+    LocateReplyMessage_1_0(ORB orb)
+    {
     }
 
-    LocateReplyMessage_1_0(ORB orb, int _request_id,
-            int _locate_status, IOR _ior) {
+    LocateReplyMessage_1_0(ORB orb, int _request_id, int _locate_status, IOR _ior)
+    {
         super(Message.GIOPBigMagic, false, Message.GIOPLocateReply, 0);
         request_id = _request_id;
         locate_status = _locate_status;
@@ -64,44 +67,51 @@ public final class LocateReplyMessage_1_0 extends Message_1_0
 
     // Accessor methods
 
-    public int getRequestId() {
+    public int getRequestId()
+    {
         return this.request_id;
     }
 
-    public int getReplyStatus() {
+    public int getReplyStatus()
+    {
         return this.locate_status;
     }
 
-    public short getAddrDisposition() {
+    public short getAddrDisposition()
+    {
         return KeyAddr.value;
     }
 
-    public SystemException getSystemException(String message) {
-        return null;  // 1.0 LocateReply body does not contain SystemException
+    public SystemException getSystemException(String message)
+    {
+        return null; // 1.0 LocateReply body does not contain SystemException
     }
 
-    public IOR getIOR() {
+    public IOR getIOR()
+    {
         return this.ior;
     }
 
     // IO methods
 
-    public void read(org.omg.CORBA.portable.InputStream istream) {
+    public void read(org.omg.CORBA.portable.InputStream istream)
+    {
         super.read(istream);
         this.request_id = istream.read_ulong();
         this.locate_status = istream.read_long();
         isValidReplyStatus(this.locate_status); // raises exception on error
 
         // The code below reads the reply body if status is OBJECT_FORWARD
-        if (this.locate_status == OBJECT_FORWARD) {
+        if (this.locate_status == OBJECT_FORWARD)
+        {
             CDRInputStream cdr = (CDRInputStream) istream;
-            this.ior = IORFactories.makeIOR( cdr ) ;
+            this.ior = IORFactories.makeIOR(cdr);
         }
     }
 
-    // Note, this writes only the header information.
-    // IOR may be written afterwards into the reply mesg body.
-    public void write(org.omg.CORBA.portable.OutputStream ostream) {
+    // Note, this writes only the header information. IOR may be written afterwards into the reply mesg body.
+    public void write(org.omg.CORBA.portable.OutputStream ostream)
+    {
         super.write(ostream);
         ostream.write_ulong(this.request_id);
         ostream.write_long(this.locate_status);
@@ -109,21 +119,21 @@ public final class LocateReplyMessage_1_0 extends Message_1_0
 
     // Static methods
 
-    public static void isValidReplyStatus(int replyStatus) {
-        switch (replyStatus) {
-        case UNKNOWN_OBJECT :
-        case OBJECT_HERE :
-        case OBJECT_FORWARD :
-            break;
-        default :
-            ORBUtilSystemException localWrapper = ORBUtilSystemException.get(
-                CORBALogDomains.RPC_PROTOCOL ) ;
-            throw localWrapper.illegalReplyStatus( CompletionStatus.COMPLETED_MAYBE);
+    public static void isValidReplyStatus(int replyStatus)
+    {
+        switch (replyStatus)
+        {
+            case UNKNOWN_OBJECT :
+            case OBJECT_HERE :
+            case OBJECT_FORWARD :
+                break;
+            default :
+                ORBUtilSystemException localWrapper = ORBUtilSystemException.get(CORBALogDomains.RPC_PROTOCOL);
+                throw localWrapper.illegalReplyStatus(CompletionStatus.COMPLETED_MAYBE);
         }
     }
 
-    public void callback(MessageHandler handler)
-        throws java.io.IOException
+    public void callback(MessageHandler handler) throws java.io.IOException
     {
         handler.handleInput(this);
     }
